requirements.txt
django
djangorestframework
django-cors-headers
Pillow
drf-yasg
django-modeltranslation
django-filter